Fix task stack corruption when making sub task
I thought that `Box::new([0; size])` was special somehow, and the array
was allocated directly on the heap. Nope, it was allocated on the stack.
This is a guaranteed way to overflow the stack when we are running a
task because we are creating an array that is larger than our current
stack size.